cbt-exercise-generator-service/requirements.txt
fastapi
uvicorn
numpy
stable-baselines3
torch
transformers
google-cloud-firestore
dotenv
google.generativeai
firebase_admin
pydantic[email]